Trait s2n_quic_platform::message::Message

pub trait Message {
    type Handle: Handle;

    const SUPPORTS_GSO: bool;

    fn ecn(&self) -> ExplicitCongestionNotification;
    fn set_ecn(
        &mut self,
        ecn: ExplicitCongestionNotification,
        remote_address: &SocketAddress
    );
    fn remote_address(&self) -> Option<SocketAddress>;
    fn set_remote_address(&mut self, remote_address: &SocketAddress);
    fn path_handle(&self) -> Option<Self::Handle>;
    fn payload_len(&self) -> usize;
    unsafe fn set_payload_len(&mut self, payload_len: usize);
    fn can_gso<M: Message<Handle = Self::Handle>>(&self, other: &mut M) -> bool;
    fn replicate_fields_from(&mut self, other: &Self);
    fn payload_ptr(&self) -> *const u8;
    fn payload_ptr_mut(&mut self) -> *mut u8;
    unsafe fn reset(&mut self, mtu: usize);

    fn payload(&self) -> &[u8]  { ... }
    fn payload_mut(&mut self) -> &mut [u8]  { ... }
    fn set_segment_size(&mut self, _size: usize) { ... }
    fn as_ptr(&self) -> *const c_void { ... }
    fn as_mut_ptr(&mut self) -> *mut c_void { ... }

}

An abstract message that can be sent and received on a network

Required Associated Types

type Handle: Handle

Required Associated Constants

const SUPPORTS_GSO: bool

Required Methods

fn ecn(&self) -> ExplicitCongestionNotification

Returns the ECN values for the message

fn set_ecn(
&mut self,
ecn: ExplicitCongestionNotification,
remote_address: &SocketAddress
)

Sets the ECN values for the message

fn remote_address(&self) -> Option<SocketAddress>

Returns the SocketAddress for the message

fn set_remote_address(&mut self, remote_address: &SocketAddress)

Sets the SocketAddress for the message

fn path_handle(&self) -> Option<Self::Handle>

Returns the path handle for the message

fn payload_len(&self) -> usize

Returns the length of the payload

unsafe fn set_payload_len(&mut self, payload_len: usize)

Sets the payload length for the message

Safety

This method should only set the payload less than or equal to its initially allocated size.

fn can_gso<M: Message<Handle = Self::Handle>>(&self, other: &mut M) -> bool

Returns true if this message can be included in the same GSO payload as the other message

fn replicate_fields_from(&mut self, other: &Self)

Copies the relevant fields inside of one message into another.

Panics

This should used in scenarios where the data pointers are the same.

fn payload_ptr(&self) -> *const u8

Returns a pointer for the message payload

fn payload_ptr_mut(&mut self) -> *mut u8

Returns a mutable pointer for the message payload

unsafe fn reset(&mut self, mtu: usize)

Resets the message for future use

Safety

This method should only set the MTU to the original value

Provided Methods

fn payload(&self) -> &[u8]

Returns a slice for the message payload

fn payload_mut(&mut self) -> &mut [u8]

Returns a mutable slice for the message payload

fn set_segment_size(&mut self, _size: usize)

Sets the segment size for the message payload

fn as_ptr(&self) -> *const c_void

Returns a pointer to the Message

fn as_mut_ptr(&mut self) -> *mut c_void

Returns a mutable pointer to the Message