Trait s2n_quic_platform::message::Ring

pub trait Ring {
    type Message: Message;

    fn len(&self) -> usize;
    fn mtu(&self) -> usize;
    fn max_gso(&self) -> usize;
    fn disable_gso(&mut self);
    fn as_slice(&self) -> &[Self::Message];
    fn as_mut_slice(&mut self) -> &mut [Self::Message];

    fn is_empty(&self) -> bool { ... }
}

A message ring used to back a queue

Required Associated Types

type Message: Message

The type of message that is stored in the ring

Required Methods

fn len(&self) -> usize

Returns the length of the ring

This value should be half the length of the slice returned to ensure contiguous access.

fn mtu(&self) -> usize

Returns the maximum transmission unit for the ring

fn max_gso(&self) -> usize

Returns the maximum number of GSO segments that can be used

fn disable_gso(&mut self)

Disables the ability for the ring to send GSO messages

This will be called in case the runtime encounters an IO error and will try again with GSO disabled.

fn as_slice(&self) -> &[Self::Message]

Returns all of the messages in the ring

The first half of the slice should be duplicated into the second half

fn as_mut_slice(&mut self) -> &mut [Self::Message]

Returns a mutable slice of the messages in the ring

Provided Methods

fn is_empty(&self) -> bool

Returns true if the ring is empty